完成活动部分接口

+++ gmh-web/src/main/java/com/zes/squad/gmh/web/entity/vo/ActivityContentVo.java
@@ -1,15 +1,17 @@
 package com.zes.squad.gmh.web.entity.vo;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class ActivityContentVo {
 
-    private Long    id;
-    private String  type;
-    private String  relatedId;
-    private String  content;
-    private Integer number;
-    private String  remark;
+    private Long       id;
+    private String     type;
+    private String     relatedId;
+    private String     content;
+    private BigDecimal number;
+    private String     remark;
 
 }
